+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added `extension` to GraphQL type
 - Added `format` to GraphQL type
 - Added `formatForCountry` to GraphQL type
+- Added `formatForMobileDialing` to GraphQL type
 - Added `timezones` to GraphQL type
 - Added support for `locale` and `region` arguments to `description` GraphQL field